28.3 Finished-all 28

+++ src/products/products.controller.ts
@@ -1,6 +1,9 @@
-import { Controller, Get, Param, Delete, Body, Post } from '@nestjs/common';
+import { Controller, Get, Param, Delete, Body, Post, Put } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { ParseUUIDPipe } from '@nestjs/common';
+import { UpdateProductDTO } from './dtos/update-product.dto';
+import { NotFoundException } from '@nestjs/common';
 
 @Controller('products')
 export class ProductsController {
@@ -12,18 +15,35 @@ export class ProductsController {
     }
 
     @Get('/:id')
-    public getById(@Param('id') id: string) {
-        return this.productsService.getById(id);
-    }
+    getById(@Param('id', new ParseUUIDPipe()) id: string) {
+        const prod = this.productsService.getById(id);
+        if (!prod) throw new NotFoundException('Product not found');
+        return prod;
+    }   
 
     @Delete('/:id')
-    deleteById(@Param('id') id: string) {
+    deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+        if (!this.productsService.getById(id))
+            throw new NotFoundException('Product not found');
         this.productsService.deleteById(id);
         return { success: true };
     }
-    
+
     @Post('/')
     create(@Body() productData: CreateProductDTO) {
         return this.productsService.create(productData);
     }   
+
+    @Put('/:id')
+    updateById(
+        @Param('id', new ParseUUIDPipe()) id: string,
+        @Body() productData: UpdateProductDTO,
+    ) {
+        if (!this.productsService.getById(id))
+        throw new NotFoundException('Product not found');
+
+        this.productsService.updateById(id, productData);
+        return { success: true };
+    }
+
 }
